Features/char customizer (#34)
* LeftUI

CharCustomizer is sliding in when clicking on Player. CC is hiding when hiting the closeButton. PlayerModel is seperated in body parts.

* Character Customizer is working with test Textures

Character Customizer has a task bar where you can choose between different parts of the body. By clicking on the clothes the look of the player is changing.

* NavButtons added

Added 2 Buttons that can navigate through the items. Also changed the classnames and general overlay of the left UI.

* Changed TabBar

Added Background to TabBarButtons.

* Added Textures and BugFixes

New textures for pants, faces and hats. Added ExceptionHandler to CurrentTab function.

* fix shop animation

* added Customizer to the right UI

* Fixed that customizable items are out of the box

* Revert "Fixed that customizable items are out of the box"

This reverts commit a93923b263de6b731fb08a795d94c7e3d201dc47.

* Fix CharCustomizer

fix scaling issue

Co-authored-by: Justus <Justernst.je@gmail.com>
Co-authored-by: Justus-e <56258792+Justus-e@users.noreply.github.com>